Track
Made a track as the background using tiles from an assets package (https://darkcesium.itch.io/pixel-roads-free-pack-1). The particular track is made in a function so that is doesn't make the game loop messy. As well as there being another function to set the images and scale them the the size of the window. 
The track will only work at the moment with this current window size and the tiles position are set relative to the window size.